refactor: alert component
