+++ src/components/ui/Alert/Alert.tsx
@@ -1,88 +1,67 @@
-import React from 'react';
-
-import { AlertIcon } from './alertIcon';
+import { AlertProps } from './types';
 import { cn } from '../../../utils/cn';
+import { AlertIcon } from './alertIcon';
+import { AlertTypeClasses } from './constant';
 
-export type AlertType = 'success' | 'error' | 'info' | 'warning' | 'default';
+const Alert = ({ ...props }: AlertProps) => {
+  const {
+    title,
+    link,
+    children,
+    className,
+    alertIcon,
+    description,
+    setOpenState,
+    type = 'default',
+    openState = false,
+  } = props;
 
-const defaultBttnClass = 'py-4 px-6 border-l-4 w-full rounded';
+  if (!openState) return null;
 
-const AlertTypeClasses: Record<AlertType, string> = {
-  success: 'bg-green-50 border-l-green-600',
-  error: 'bg-warning-50 border-l-warning-600',
-  info: 'bg-blue-50 border-l-blue-600',
-  warning: 'bg-yellow-50 border-l-yellow-600',
-  default: 'bg-gray-50 border-l-gray-600',
-};
-
-export interface AlertProps {
-  type: AlertType;
-  title?: string;
-  description?: string;
-  children?: React.ReactNode;
-  link?: string;
-  openState: boolean;
-  className?: string;
-  alertIcon?: React.ReactNode;
-  setOpenState: React.Dispatch<React.SetStateAction<boolean>>;
-}
-const Alert: React.FC<AlertProps> = ({
-  type = 'default',
-  openState = false,
-  title,
-  description,
-  link,
-  className,
-  children,
-  alertIcon,
-  setOpenState,
-}: AlertProps) => {
   return (
-    <>
-      {openState === true ? (
-        <div
-          role='alert'
-          className={cn(defaultBttnClass, AlertTypeClasses[type], className)}
-        >
-          <div className='relative pl-14'>
-            <span className='icon absolute left-0 top-0 block'>
-              {alertIcon ? alertIcon : AlertIcon(type)}
+    <div
+      role='alert'
+      className={cn(
+        'w-full rounded border-l-4 px-6 py-4',
+        AlertTypeClasses[type],
+        className,
+      )}
+    >
+      <div className='relative pl-14'>
+        <span className='icon absolute left-0 top-0 block'>
+          {alertIcon ? alertIcon : AlertIcon(type)}
+        </span>
+        <div className='flex items-start justify-between gap-2'>
+          <div className='content'>
+            {title && (
+              <h3 className='title m-0 mb-1 text-base font-semibold'>
+                {title}
+              </h3>
+            )}
+            {description && (
+              <div className='description text-base font-normal'>
+                {description}
+              </div>
+            )}
+          </div>
+          <div className='flex items-center gap-1.5'>
+            <a
+              href={link}
+              className='text-xs'
+            >
+              View Details
+            </a>
+            <span
+              className='text-xs'
+              onClick={() => setOpenState(false)}
+            >
+              Dismiss
             </span>
-            <div className='flex items-start justify-between gap-2'>
-              <div className='content'>
-                {title && (
-                  <h3 className='title m-0 mb-1 text-base font-semibold'>
-                    {title}
-                  </h3>
-                )}
-                {description && (
-                  <div className='description text-base font-normal'>
-                    {description}
-                  </div>
-                )}
-              </div>
-              <div className='flex items-center gap-1.5'>
-                <a
-                  href={link}
-                  className='text-xs'
-                >
-                  View Details
-                </a>
-                <span
-                  className='text-xs'
-                  onClick={() => setOpenState(false)}
-                >
-                  Dismiss
-                </span>
-              </div>
-            </div>
-            {children && <div>{children}</div>}
           </div>
         </div>
-      ) : (
-        <></>
-      )}
-    </>
+        {children && <div>{children}</div>}
+      </div>
+    </div>
   );
 };
