+++ src/components/ui/Alert/constant/index.tsx
@@ -1,4 +1,4 @@
-import { AlertType } from '../Alert';
+import { AlertType } from '../types';
 
 export const mockData: {
   link?: string;
@@ -14,4 +14,12 @@ export const mockData: {
   link: 'https://example.com',
   description: 'Test Description',
   children: <div data-testid='test-child'>Child Content</div>,
+};
+
+export const AlertTypeClasses: Record<AlertType, string> = {
+  success: 'bg-green-50 border-l-green-600',
+  error: 'bg-warning-50 border-l-warning-600',
+  info: 'bg-blue-50 border-l-blue-600',
+  warning: 'bg-yellow-50 border-l-yellow-600',
+  default: 'bg-gray-50 border-l-gray-600',
 };
